feat(tui): add System tab, fix address balance overflow, adjust verify tolerances
System tab (new):
- 3rd TUI tab showing system environment, store paths, and RocksDB parameters
- Live indexer mode (Normal/Bulk Sync) and WBM usage from Redis memory stats
- Responsive layout: 3-column (wide) or stacked (compact/narrow terminals)
- Correct per-CF L0 values (12/24 normal, 64/128 bulk) instead of DB-level defaults
- Direct I/O display reads actual CKBADGER_DIRECT_IO_READS env var
- Tab bar hides sync-specific indicators ([Compact]/[Diag:X]) on System tab
- Help popup lists all three tab names

Address balance precision:
- Widen balance/occupied deltas from i64 to i128 in address balance change maps
- Prevents intermediate overflow when summing large capacity values across
  many inputs/outputs within a single batch

Verify tolerance adjustments:
- Bump avg_hash_rate and avg_difficulty tolerance from 0.07 to 0.08
- Bump circulation_ratio tolerance from 0.012 to 0.02
- Fix knowledge_size comparison: convert CKB units back to shannons

+++ crates/indexer/tests/pipeline_consistency.rs
@@ -354,7 +354,7 @@ fn test_address_balance_update_receive() {
     let lock_hash = vec![0xAAu8; 32];
     let tx_hash = vec![0x01u8; 32];
 
-    let changes: HashMap<Vec<u8>, (i64, i32, i32, i64, i64, &[u8], i64)> = [(
+    let changes: HashMap<Vec<u8>, (i128, i32, i32, i64, i64, &[u8], i128)> = [(
         lock_hash.clone(),
         (100_00000000, 1, 1, 1, 1000, tx_hash.as_slice(), 0),
     )]
@@ -382,7 +382,7 @@ fn test_address_balance_update_send() {
     let tx_hash1 = vec![0x01u8; 32];
     let tx_hash2 = vec![0x02u8; 32];
 
-    let receive: HashMap<Vec<u8>, (i64, i32, i32, i64, i64, &[u8], i64)> = [(
+    let receive: HashMap<Vec<u8>, (i128, i32, i32, i64, i64, &[u8], i128)> = [(
         lock_hash.clone(),
         (100_00000000, 1, 1, 1, 1000, tx_hash1.as_slice(), 0),
     )]
@@ -394,7 +394,7 @@ fn test_address_balance_update_send() {
         .unwrap();
     batch.commit().unwrap();
 
-    let send: HashMap<Vec<u8>, (i64, i32, i32, i64, i64, &[u8], i64)> = [(
+    let send: HashMap<Vec<u8>, (i128, i32, i32, i64, i64, &[u8], i128)> = [(
         lock_hash.clone(),
         (-30_00000000, 0, 1, 1, 2000, tx_hash2.as_slice(), 0),
     )]
@@ -421,7 +421,7 @@ fn test_address_balance_occupied_delta_applied() {
     let tx_hash2 = vec![0x02u8; 32];
 
     // Receive: creates a cell with 6100 CKB occupied
-    let receive: HashMap<Vec<u8>, (i64, i32, i32, i64, i64, &[u8], i64)> = [(
+    let receive: HashMap<Vec<u8>, (i128, i32, i32, i64, i64, &[u8], i128)> = [(
         lock_hash.clone(),
         (
             100_00000000,
@@ -445,7 +445,7 @@ fn test_address_balance_occupied_delta_applied() {
     assert_eq!(balance.occupied_capacity, 6100_00000000);
 
     // Consume old cell (-6100) and create new smaller cell (+4100)
-    let update: HashMap<Vec<u8>, (i64, i32, i32, i64, i64, &[u8], i64)> = [(
+    let update: HashMap<Vec<u8>, (i128, i32, i32, i64, i64, &[u8], i128)> = [(
         lock_hash.clone(),
         (0, 0, 1, 1, 2000, tx_hash2.as_slice(), -2000_00000000),
     )]
@@ -469,7 +469,7 @@ fn test_address_balance_occupied_underflow_errors() {
 
     // Apply a negative occupied_delta larger than what exists (0)
     // Should fail fast instead of silently clamping.
-    let changes: HashMap<Vec<u8>, (i64, i32, i32, i64, i64, &[u8], i64)> = [(
+    let changes: HashMap<Vec<u8>, (i128, i32, i32, i64, i64, &[u8], i128)> = [(
         lock_hash.clone(),
         (
             100_00000000,
